requirements.txt
Django>=4.2
python-dotenv>=1.0.0
youtube-transcript-api== 1.2.2